+++ src/main/java/de/christophlorenz/jweatherflux/config/ForwarderProperties.java
@@ -1,5 +1,6 @@
 package de.christophlorenz.jweatherflux.config;
 
+import de.christophlorenz.jweatherflux.forwarder.WetterCom;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 
@@ -9,6 +10,7 @@ public class ForwarderProperties {
 
   private TestingForwarder testingForwarder;
   private Wunderground wunderground;
+  private WetterCom wetterCom;
 
   public void setTestingForwarder(TestingForwarder testingForwarder) {
     this.testingForwarder = testingForwarder;
@@ -27,11 +29,20 @@ public class ForwarderProperties {
     this.wunderground = wunderground;
   }
 
+  public WetterCom getWetterCom() {
+    return wetterCom;
+  }
+
+  public void setWetterCom(WetterCom wetterCom) {
+    this.wetterCom = wetterCom;
+  }
+
   @Override
   public String toString() {
     return "ForwarderProperties{" +
         "testingForwarder=" + testingForwarder +
         ", wunderground=" + wunderground +
+        ", wetterCom=" + wetterCom +
         '}';
   }
 
@@ -54,10 +65,15 @@ public class ForwarderProperties {
     }
   }
 
+  protected static String mask(String stringToMask) {
+    return stringToMask.replaceAll(".", "*");
+  }
+
   public static class Wunderground {
     private String url;
     private String id;
     private String key;
+
     private int period;
 
     public String getUrl() {
@@ -97,8 +113,57 @@ public class ForwarderProperties {
       return "Wunderground{" +
           "url='" + url + '\'' +
           ", id='" + id + '\'' +
-          ", key='" + key + '\'' +
+          ", key='" + mask(key) + '\'' +
           ", period=" + period +
+          '}';
+    }
+  }
+
+  public static class WetterCom {
+    private String url;
+    private String userId;
+    private String idKennwort;
+    private int interval;
+
+    public String getUrl() {
+      return url;
+    }
+
+    public void setUrl(String url) {
+      this.url = url;
+    }
+
+    public String getUserId() {
+      return userId;
+    }
+
+    public void setUserId(String userId) {
+      this.userId = userId;
+    }
+
+    public String getIdKennwort() {
+      return idKennwort;
+    }
+
+    public void setIdKennwort(String idKennwort) {
+      this.idKennwort = idKennwort;
+    }
+
+    public int getInterval() {
+      return interval;
+    }
+
+    public void setInterval(int interval) {
+      this.interval = interval;
+    }
+
+    @Override
+    public String toString() {
+      return "WetterCom{" +
+          "url='" + url + '\'' +
+          "userId='" + userId + '\'' +
+          ", idKennwort='" + mask(idKennwort) + '\'' +
+          ", interval=" + interval +
           '}';
     }
   }