+++ src/main/java/de/christophlorenz/jweatherflux/forwarder/AuthorizationException.java
@@ -1,0 +1,9 @@
+package de.christophlorenz.jweatherflux.forwarder;
+
+public class AuthorizationException extends
+    Exception {
+
+  public AuthorizationException(String msg) {
+    super(msg);
+  }
+}