+++ src/main/java/de/christophlorenz/jweatherflux/forwarder/InvalidDataException.java
@@ -1,9 +1,13 @@
 package de.christophlorenz.jweatherflux.forwarder;
 
 public class InvalidDataException extends
-    Throwable {
+    Exception {
 
   public InvalidDataException(String msg) {
     super(msg);
+  }
+
+  public InvalidDataException(String msg, Exception e) {
+    super(msg, e);
   }
 }