+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased]
 
+## [0.1.0-SNAPSHOT]() - 2022-02-18
+
+### Added
+
+- Added wetter.com forwarder
+
+### Changed
+
+- Starting forwarder refactoring
+
 ## [0.0.4-SNAPSHOT]() - 2022-02-10
 
 ### Added